uploading data to the map

+++ ex00/BitcoinExchange.hpp
@@ -5,4 +5,4 @@
 #include <fstream>
 
 void parseTheFile(std::string file);
-void loadData(std::map<unsigned long, unsigned int>& mydb);
+void loadData(std::map<unsigned long, double>& mydb);
